# apps/web/app/api/admin/guests/[id]/route.ts
import { currentUser } from "@clerk/nextjs/server";
import { type NextRequest, NextResponse } from "next/server";
import { env } from "@/env";
import { db } from "@/lib/db";

/**
 * Helper function to delete a party if it has no guests remaining
 */
async function deleteEmptyParty(partyId: string): Promise<void> {
  try {
    const guestCount = await db
      .selectFrom("guests")
      .select(db.fn.count("id").as("count"))
      .where("party_id", "=", partyId)
      .executeTakeFirst();

    if (guestCount && Number(guestCount.count) === 0) {
      await db.deleteFrom("parties").where("id", "=", partyId).execute();
    }
  } catch (error) {
    // Log but don't fail the main operation
    console.error("Error cleaning up empty party:", error);
  }
}

/**
 * Get a guest by ID
 * @description Get a guest and their plus-one if they have one (admin only)
 * @pathParams IdParams
 * @response 200:GuestDetailResponse
 * @auth bearer
 * @tag Admin - Guests
 * @openapi
 */
export async function GET(
  _request: NextRequest,
  { params }: { params: Promise<{ id: string }> },
) {
  try {
    const user = await currentUser();

    if (!user) {
      return NextResponse.json({ error: "Unauthorized" }, { status: 401 });
    }

    // Check if user is admin
    const adminEmails = env.ADMIN_EMAILS?.split(",").map((e) =>
      e.trim().toLowerCase(),
    );
    const userEmail = user.emailAddresses[0]?.emailAddress?.toLowerCase();

    if (!adminEmails?.includes(userEmail || "")) {
      return NextResponse.json({ error: "Forbidden" }, { status: 403 });
    }

    const { id } = await params;

    // Fetch the guest
    const guest = await db
      .selectFrom("guests")
      .selectAll()
      .where("id", "=", id)
      .executeTakeFirst();

    if (!guest) {
      return NextResponse.json({ error: "Guest not found" }, { status: 404 });
    }

    // Fetch plus-one if exists
    const plusOne = await db
      .selectFrom("guests")
      .selectAll()
      .where("primary_guest_id", "=", id)
      .where("is_plus_one", "=", true)
      .executeTakeFirst();

    return NextResponse.json({ guest, plusOne: plusOne || null });
  } catch (error) {
    console.error("Error in GET /api/admin/guests/[id]:", error);
    return NextResponse.json(
      { error: "Internal server error" },
      { status: 500 },
    );
  }
}

/**
 * Update a guest
 * @description Update a guest's details, manage plus-ones, and handle party reassignment (admin only)
 * @pathParams IdParams
 * @body UpdateGuestBody
 * @response 200:UpdateGuestResponse
 * @auth bearer
 * @tag Admin - Guests
 * @openapi
 */
export async function PATCH(
  request: NextRequest,
  { params }: { params: Promise<{ id: string }> },
) {
  try {
    const user = await currentUser();

    if (!user) {
      return NextResponse.json({ error: "Unauthorized" }, { status: 401 });
    }

    // Check if user is admin
    const adminEmails = env.ADMIN_EMAILS?.split(",").map((e) =>
      e.trim().toLowerCase(),
    );
    const userEmail = user.emailAddresses[0]?.emailAddress?.toLowerCase();

    if (!adminEmails?.includes(userEmail || "")) {
      return NextResponse.json({ error: "Forbidden" }, { status: 403 });
    }

    const { id } = await params;
    const body = await request.json();
    const {
      firstName,
      lastName,
      email,
      side,
      list,
      plusOneAllowed,
      plusOneFirstName,
      plusOneLastName,
      mailingAddress,
      physicalInviteSent,
      phoneNumber,
      whatsapp,
      preferredContactMethod,
      family,
      under21,
      threeAndUnder,
      notes,
      gender,
      bridalPartyRole,
      partyId,
    } = body;

    // Kysely query - fetch the current guest to check if they have a plus one
    const currentGuest = await db
      .selectFrom("guests")
      .selectAll()
      .where("id", "=", id)
      .executeTakeFirst();

    if (!currentGuest) {
      return NextResponse.json({ error: "Guest not found" }, { status: 404 });
    }

    // Handle party change - get the new party's invite_code if partyId changed
    let newPartyId = currentGuest.party_id;
    let newInviteCode = currentGuest.invite_code;
    const sourcePartyId = currentGuest.party_id; // Track for cleanup

    if (partyId !== undefined && partyId !== currentGuest.party_id) {
      if (partyId) {
        // Moving to an existing party
        const targetParty = await db
          .selectFrom("parties")
          .select(["id", "invite_code"])
          .where("id", "=", partyId)
          .executeTakeFirst();

        if (!targetParty) {
          return NextResponse.json(
            { error: "Target party not found" },
            { status: 404 },
          );
        }

        newPartyId = targetParty.id;
        newInviteCode = targetParty.invite_code;
      } else {
        // Removing from party (set to null)
        newPartyId = null;
      }
    }

    // Update the primary guest
    const updatedGuest = await db
      .updateTable("guests")
      .set({
        first_name: firstName || currentGuest.first_name,
        last_name:
          lastName !== undefined ? lastName || null : currentGuest.last_name,
        email: email || null,
        side: side !== undefined ? side : currentGuest.side,
        list: list || currentGuest.list,
        plus_one_allowed: plusOneAllowed || false,
        mailing_address:
          mailingAddress !== undefined
            ? mailingAddress || null
            : currentGuest.mailing_address,
        physical_invite_sent:
          physicalInviteSent !== undefined
            ? physicalInviteSent
            : currentGuest.physical_invite_sent,
        phone_number:
          phoneNumber !== undefined
            ? phoneNumber || null
            : currentGuest.phone_number,
        whatsapp:
          whatsapp !== undefined ? whatsapp || null : currentGuest.whatsapp,
        preferred_contact_method:
          preferredContactMethod !== undefined
            ? preferredContactMethod || null
            : currentGuest.preferred_contact_method,
        family: family !== undefined ? family : currentGuest.family,
        under_21: under21 !== undefined ? under21 : currentGuest.under_21,
        three_and_under:
          threeAndUnder !== undefined
            ? threeAndUnder
            : currentGuest.three_and_under,
        notes: notes !== undefined ? notes || null : currentGuest.notes,
        gender: gender !== undefined ? gender || null : currentGuest.gender,
        bridal_party_role:
          bridalPartyRole !== undefined
            ? bridalPartyRole || null
            : currentGuest.bridal_party_role,
        party_id: newPartyId,
        invite_code: newInviteCode,
      })
      .where("id", "=", id)
      .returningAll()
      .executeTakeFirst();

    // Handle plus one logic
    if (plusOneAllowed) {
      // Check if plus one already exists
      const existingPlusOne = await db
        .selectFrom("guests")
        .selectAll()
        .where("primary_guest_id", "=", id)
        .where("is_plus_one", "=", true)
        .executeTakeFirst();

      // Determine plus-one name: use provided name or create placeholder
      const primaryFullName =
        `${updatedGuest?.first_name || currentGuest.first_name}${
          updatedGuest?.last_name || currentGuest.last_name
            ? ` ${updatedGuest?.last_name || currentGuest.last_name}`
            : ""
        }`.trim();
      const plusOneFirstNameFinal = plusOneFirstName?.trim() || primaryFullName;
      const plusOneLastNameFinal = plusOneFirstName?.trim()
        ? plusOneLastName || null
        : "- Plus One";

      if (existingPlusOne) {
        // Update existing plus one
        // Note: list and family are automatically cascaded by database trigger
        await db
          .updateTable("guests")
          .set({
            first_name: plusOneFirstNameFinal,
            last_name: plusOneLastNameFinal,
            side: side !== undefined ? side : currentGuest.side,
            party_id: newPartyId,
            invite_code: newInviteCode,
          })
          .where("id", "=", existingPlusOne.id)
          .execute();
      } else {
        // Create new plus one
        // Note: list and family will be automatically set by database trigger on first update
        // For initial creation, we set them explicitly since trigger only runs on UPDATE
        await db
          .insertInto("guests")
          .values({
            first_name: plusOneFirstNameFinal,
            last_name: plusOneLastNameFinal,
            email: null, // Plus-ones don't have their own email
            invite_code: newInviteCode,
            party_id: newPartyId,
            side: side !== undefined ? side : currentGuest.side,
            list: list || currentGuest.list,
            is_plus_one: true,
            plus_one_allowed: false, // Plus ones themselves don't get plus ones
            primary_guest_id: id,
            rsvp_status: "pending",
            number_of_resends: 0,
            physical_invite_sent: false,
            family: family !== undefined ? family : currentGuest.family,
            under_21: under21 !== undefined ? under21 : currentGuest.under_21,
            three_and_under:
              threeAndUnder !== undefined
                ? threeAndUnder
                : currentGuest.three_and_under,
          })
          .execute();
      }
    } else {
      // Remove plus one if it exists and plusOneAllowed is false
      await db
        .deleteFrom("guests")
        .where("primary_guest_id", "=", id)
        .where("is_plus_one", "=", true)
        .execute();
    }

    // Clean up the source party if it's now empty (guest was moved to a different party)
    if (sourcePartyId && sourcePartyId !== newPartyId) {
      await deleteEmptyParty(sourcePartyId);
    }

    return NextResponse.json({ guest: updatedGuest });
  } catch (error) {
    console.error("Error in PATCH /api/admin/guests/[id]:", error);
    return NextResponse.json(
      { error: "Internal server error" },
      { status: 500 },
    );
  }
}