# apps/web/public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Wedding Website API",
    "version": "1.0.0",
    "description": "API for the wedding website - manage guests, events, RSVPs, gifts, photos, and more"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
  "paths": {
...
    "/admin/guests/{id}": {
      "get": {
        "operationId": "get-admin-guests-{id}",
        "summary": "Get a guest by ID",
        "description": "Get a guest and their plus-one if they have one (admin only)",
        "tags": ["Admin - Guests"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string",
              "description": "Resource UUID"
            },
            "required": true,
            "description": "Resource UUID",
            "example": "123"
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GuestDetailResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "operationId": "patch-admin-guests-{id}",
        "summary": "Update a guest",
        "description": "Update a guest's details, manage plus-ones, and handle party reassignment (admin only)",
        "tags": ["Admin - Guests"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string",
              "description": "Resource UUID"
            },
            "required": true,
            "description": "Resource UUID",
            "example": "123"
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateGuestBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateGuestResponse"
                }
              }
            }
          }
        }
      }
    },
    "/admin/guests/bulk-send-email": {
      "post": {
        "operationId": "post-admin-guests-bulk-send-email",
        "summary": "Bulk send invitation emails",
        "description": "Send wedding invitation emails to multiple guests at once using the default or a custom Resend template",
        "tags": ["Admin - Guests"],
        "parameters": [],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkSendEmailBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "Whether the operation succeeded"
          }
        },
        "required": ["success"]
      },
...
      "GuestDetailResponse": {
        "type": "object",
        "properties": {
          "guest": {
            "allOf": [
              {
                "$ref": "#/components/schemas/GuestResponse"
              }
            ],
            "description": "Primary guest"
          },
          "plusOne": {
            "allOf": [
              {
                "$ref": "#/components/schemas/GuestResponse"
              }
            ],
            "description": "Plus-one guest if exists"
          }
        },
        "required": ["guest", "plusOne"]
      },
      "UpdateGuestBody": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "description": "First name"
          },
          "lastName": {
            "type": "string",
            "description": "Last name"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "Email address"
          },
          "side": {
            "type": "string",
            "enum": ["bride", "groom", "both"],
            "description": "Side"
          },
          "list": {
            "type": "string",
            "enum": ["a", "b", "c"],
            "description": "Guest list tier"
          },
          "plusOneAllowed": {
            "type": "boolean",
            "description": "Whether a plus-one is allowed"
          },
          "plusOneFirstName": {
            "type": "string",
            "description": "Plus-one first name"
          },
          "plusOneLastName": {
            "type": "string",
            "description": "Plus-one last name"
          },
          "mailingAddress": {
            "type": "string",
            "description": "Mailing address"
          },
          "physicalInviteSent": {
            "type": "boolean",
            "description": "Whether physical invite was sent"
          },
          "phoneNumber": {
            "type": "string",
            "description": "Phone number"
          },
          "whatsapp": {
            "type": "string",
            "description": "WhatsApp number"
          },
          "preferredContactMethod": {
            "type": "string",
            "enum": ["email", "text", "whatsapp", "phone_call"],
            "description": "Preferred contact method"
          },
          "family": {
            "type": "boolean",
            "description": "Family group"
          },
          "under21": {
            "type": "boolean",
            "description": "Under 21"
          },
          "threeAndUnder": {
            "type": "boolean",
            "description": "3 and under"
          },
          "notes": {
            "type": "string",
            "description": "Admin notes"
          },
          "gender": {
            "type": "string",
            "enum": ["male", "female"],
            "description": "Gender"
          },
          "bridalPartyRole": {
            "type": "string",
            "enum": ["groomsman", "best_man", "bridesmaid", "maid_of_honor"],
            "description": "Bridal party role"
          },
          "partyId": {
            "type": "string",
            "format": "uuid",
            "description": "Party ID to move guest to"
          }
        }
      },
      "UpdateGuestResponse": {
        "type": "object",
        "properties": {
          "guest": {
            "allOf": [
              {
                "$ref": "#/components/schemas/GuestResponse"
              }
            ],
            "description": "Updated guest"
          }
        },
        "required": ["guest"]
      },
      "BulkSendEmailBody": {},
...
      "GuestResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Guest UUID"
          },
          "first_name": {
            "type": "string",
            "description": "First name"
          },
          "last_name": {
            "type": "string",
            "nullable": true,
            "description": "Last name"
          },
          "email": {
            "type": "string",
            "format": "email",
            "nullable": true,
            "description": "Email address"
          },
          "invite_code": {
            "type": "string",
            "description": "Unique invite code"
          },
          "rsvp_status": {
            "type": "string",
            "enum": ["pending", "yes", "no"],
            "description": "RSVP status"
          },
          "plus_one_allowed": {
            "type": "boolean",
            "description": "Whether a plus-one is allowed"
          },
          "dietary_restrictions": {
            "type": "string",
            "nullable": true,
            "description": "Dietary restrictions"
          },
          "side": {
            "type": "string",
            "enum": ["bride", "groom", "both"],
            "nullable": true,
            "description": "Bride or groom side"
          },
          "list": {
            "type": "string",
            "enum": ["a", "b", "c"],
            "description": "Guest list tier"
          },
          "is_plus_one": {
            "type": "boolean",
            "description": "Whether this guest is a plus-one"
          },
          "primary_guest_id": {
            "type": "string",
            "format": "uuid",
            "nullable": true,
            "description": "ID of the primary guest if this is a plus-one"
          },
          "party_id": {
            "type": "string",
            "format": "uuid",
            "nullable": true,
            "description": "Party group ID"
          },
          "family": {
            "type": "boolean",
            "description": "Whether this is a family group"
          },
          "under_21": {
            "type": "boolean",
            "description": "Whether this guest is under 21"
          },
          "three_and_under": {
            "type": "boolean",
            "description": "Whether this guest is 3 or under"
          },
          "notes": {
            "type": "string",
            "nullable": true,
            "description": "Admin notes"
          },
          "gender": {
            "type": "string",
            "enum": ["male", "female"],
            "nullable": true,
            "description": "Gender"
          },
          "bridal_party_role": {
            "type": "string",
            "enum": ["groomsman", "best_man", "bridesmaid", "maid_of_honor"],
            "nullable": true,
            "description": "Bridal party role"
          },
          "created_at": {
            "type": "string",
            "description": "Creation timestamp"
          }
        },
        "required": [
          "id",
          "first_name",
          "last_name",
          "email",
          "invite_code",
          "rsvp_status",
          "plus_one_allowed",
          "dietary_restrictions",
          "side",
          "list",
          "is_plus_one",
          "primary_guest_id",
          "party_id",
          "family",
          "under_21",
          "three_and_under",
          "notes",
          "gender",
          "bridal_party_role",
          "created_at"
        ]
      },
...
    },
...
  }
}